requirements.txt
Adafruit-DHT
RPi.GPIO
